+++ src/pydiverse/pipedag/backend/table/sql/ddl.py
@@ -46,8 +46,8 @@ from sqlalchemy.sql.type_api import TypeEngine
 @frozen
 class Schema:
     name: str
-    prefix: str
-    suffix: str
+    prefix: str = ""
+    suffix: str = ""
 
     def get(self):
         return self.prefix + self.name + self.suffix
@@ -209,7 +209,7 @@ class RenameTable(DDLElement):
 
 
 class DropTable(DDLElement):
-    def __init__(self, name, schema: Schema, if_exists=False, cascade=False):
+    def __init__(self, name, schema: Schema | str, if_exists=False, cascade=False):
         self.name = name
         self.schema = schema
         self.if_exists = if_exists
@@ -222,7 +222,7 @@ class DropView(DDLElement):
                a 'USE <database>' statement.
     """
 
-    def __init__(self, name, schema: Schema, if_exists=False):
+    def __init__(self, name, schema: Schema | str, if_exists=False):
         self.name = name
         self.schema = schema
         self.if_exists = if_exists
@@ -233,7 +233,7 @@ class DropAlias(DDLElement):
     This is used for dialect=ibm_sa_db
     """
 
-    def __init__(self, name, schema: Schema, if_exists=False, *, engine=None):
+    def __init__(self, name, schema: Schema | str, if_exists=False, *, engine=None):
         """
         :param engine: Used if if_exists=True but the database doesn't support it.
         """
@@ -260,7 +260,7 @@ class DropProcedure(DDLElement):
                a 'USE <database>' statement.
     """
 
-    def __init__(self, name, schema: Schema, if_exists=False):
+    def __init__(self, name, schema: Schema | str, if_exists=False):
         self.name = name
         self.schema = schema
         self.if_exists = if_exists
@@ -272,7 +272,7 @@ class DropFunction(DDLElement):
                a 'USE <database>' statement.
     """
 
-    def __init__(self, name, schema: Schema, if_exists=False):
+    def __init__(self, name, schema: Schema | str, if_exists=False):
         self.name = name
         self.schema = schema
         self.if_exists = if_exists
@@ -914,8 +914,11 @@ def visit_drop_alias_ibm_db_sa(drop: DropAlias, compiler, **kw):
     if drop.if_exists:
         from pydiverse.pipedag.backend.table.sql.reflection import PipedagDB2Reflection
 
+        schema_str = (
+            drop.schema.get() if isinstance(drop.schema, Schema) else drop.schema
+        )
         if drop.name not in PipedagDB2Reflection.get_alias_names(
-            drop.engine, schema=drop.schema.get()
+            drop.engine, schema=schema_str
         ):
             return ""
         drop = DropAlias(drop.name, drop.schema, if_exists=False)
@@ -945,7 +948,8 @@ def _visit_drop_anything(
 ):
     _ = kw
     table = compiler.preparer.quote(drop.name)
-    schema = compiler.preparer.format_schema(drop.schema.get())
+    schema_str = drop.schema.get() if isinstance(drop.schema, Schema) else drop.schema
+    schema = compiler.preparer.format_schema(schema_str)
     text = [f"DROP {_type}"]
     if drop.if_exists:
         text.append("IF EXISTS")