+++ src/pydiverse/pipedag/materialize/debug.py
@@ -6,7 +6,7 @@ import structlog
 
 from pydiverse.pipedag import ConfigContext, Table
 from pydiverse.pipedag.backend import SQLTableStore
-from pydiverse.pipedag.backend.table.sql.ddl import DropTable
+from pydiverse.pipedag.backend.table.sql.ddl import DropTable, Schema
 from pydiverse.pipedag.context import TaskContext
 from pydiverse.pipedag.materialize.core import MaterializingTask
 from pydiverse.pipedag.materialize.store import mangle_table_name
@@ -15,6 +15,8 @@ from pydiverse.pipedag.util.hashing import stable_hash
 
 def materialize_table(
     table: Table,
+    config_context: ConfigContext | None = None,
+    schema: Schema | str | None = None,
     debug_suffix: str | None = None,
     flag_task_debug_tainted: bool = True,
     keep_table_name: bool = True,
@@ -29,6 +31,10 @@ def materialize_table(
     If the table name ends in %%, the %% are also replaced by a suffix.
 
     :param table: The table to be materialized.
+    :param config_context: The config context to be used for materialization.
+        If None, the current config context is used. Default: None
+    :param schema: The schema to be used for writing table. If None, the
+        schema is derived from the table's stage. Default: None
     :param debug_suffix: A suffix to be appended to the table name
         for debugging purposes. Default: None
     :param flag_task_debug_tainted: Whether to flag the task as tainted
@@ -40,13 +46,23 @@ def materialize_table(
     :param drop_if_exists: If True, try to drop the table (if exists) before recreating.
         This is only supported for SQL table stores.
     """
-    config_context = ConfigContext.get()
+    if config_context is None:
+        config_context = ConfigContext.get()
     table_store = config_context.store.table_store
 
-    task_context = TaskContext.get()
-    task: MaterializingTask = task_context.task  # type: ignore
+    task: MaterializingTask | None = None
+    if task_context := TaskContext.get():
+        task = task_context.task  # type: ignore
 
-    table.stage = task.stage
+        table.stage = task.stage
+        if schema is None:
+            schema = table_store.get_schema(table.stage.transaction_name)
+    else:
+        if schema is None:
+            raise ValueError(
+                "Parameter schema must be provided if task is not called by "
+                "normal pipedag orchestration."
+            )
 
     suffix = (
         stable_hash(str(random.randbytes(8))) + "_0000" if debug_suffix is None else ""
@@ -56,12 +72,11 @@ def materialize_table(
     if debug_suffix is not None:
         table.name += debug_suffix
 
-    if flag_task_debug_tainted:
+    if flag_task_debug_tainted and task is not None:
         task.debug_tainted = True
 
     if drop_if_exists:
         if isinstance(table_store, SQLTableStore):
-            schema = table_store.get_schema(task.stage.transaction_name)
             table_store.execute(DropTable(table.name, schema, if_exists=True))
         else:
             logger = structlog.get_logger(logger_name="Debug materialize_table")